+++ CHANGELOG.md
@@ -4,3 +4,9 @@
   * Update Navigation to use Coordinators
   * Adding "Pokemons Colors" for use latter
   * Update PokemonsViewController, improve animations, transitions and look in general
+  * Adding files
+
+## [Add, UPDATE] 2020-07-05 22:50
+  * Adding PokemonsProvider
+  * Adding petitions for get Regions and get Pokemons per region
+  * Improve some code